+++ CHANGELOG.md
@@ -12,6 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [data-service-base](https://github.com/fugerit-org/fj-service-helper-bom/blob/main/data-service-base/CHANGELOG.md)
 - [data-service-jvfs](https://github.com/fugerit-org/fj-service-helper-bom/blob/main/data-service-jvfs/CHANGELOG.md)
 
+### Added
+
+- config to properties utilities
+
+## [1.5.0 - 2025-06-30]
+
+### Added
+
+- simple-config ConfigParamsOverride
+
 ## [1.4.5 - 2024-11-19]
 
 ### Added